gstimxv4l2: Fix typecasting errors
Fixes
gstimxv4l2.c:1702:22: error: incompatible pointer to integer conversion assigning to 'unsigned long' from 'guint8 *' (aka 'unsigned char *') [-Wint-conversion]

+++ libs/v4l2_core/gstimxv4l2.c
@@ -1699,7 +1699,7 @@ gint gst_imx_v4l2_register_buffer (gpointer v4l2handle, PhyMemBlock *memblk)
   v4l2buf->type = handle->type;
   v4l2buf->memory = handle->memory_mode;
   v4l2buf->index = handle->allocated;
-  v4l2buf->m.userptr = memblk->paddr;
+  v4l2buf->m.userptr = (unsigned long)memblk->paddr;
   v4l2buf->length = memblk->size;
   handle->buffer_pair[handle->allocated].vaddr = memblk->vaddr;
